type correctness; from Ray Lai in PR 5011; ok millert@

+++ usr.bin/nc/atomicio.c
@@ -1,4 +1,4 @@
-/* $OpenBSD: atomicio.c,v 1.7 2005/05/26 01:01:08 avsm Exp $ */
+/* $OpenBSD: atomicio.c,v 1.8 2006/02/11 19:31:18 otto Exp $ */
 
 /*
  * Copyright (c) 2005 Anil Madhavapeddy.  All rights served.
@@ -57,7 +57,7 @@ atomicio(f, fd, _s, n)
 			errno = EPIPE;
 			return pos;
 		default:
-			pos += (u_int)res;
+			pos += (size_t)res;
 		}
 	}
 	return pos;
